spring: de-activate EC console input when WP is on
Spring supports Toad cable on the micro-B connector.
So let's define CONFIG_CONSOLE_RESTRICTED_INPUT on this platform
in order to de-activate the input on the EC serial console when
the Write-Protect is on.

Signed-off-by: Vincent Palatin <vpalatin@chromium.org>

BRANCH=spring
BUG=chrome-os-partner:18716
TEST=on Spring, try with and without write-protect, use successfully
the EC console in the former case, and see "Console is DISABLED" in
the latter case.

Change-Id: I843582dd4950ce235db5cbb62dad2cad1fabf72c