irq: Don't discard irqprio data due to CONFIG_LTO
The IRQ_PRIORITY macro is used directly when defining special ISRs
(commonly watchdog). Move the __keep attribute from the parent macro to
take care of those cases.

This change is probably a NOP for all platforms, since the power-on
default for the IRQ prio register seems to be zero, which is the same
priority we're setting in our direct use of IRQ_PRIORITY (for watchdog
ISRs).

BUG=chromium:634701
BRANCH=None
TEST=Verify 'prio_44' entry exists in irqprio section by checking
ec.RO.map on kevin.

Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Change-Id: Idaffc484a2ce4749c18212f179b3951ff570aed0