Update patch set 1
Patch Set 1:

Sorry, I saw this earlier, but then it had slipped past me. I think I need to understand better what the problem issue is where without 'need_static = 0' when curr.batt.is_present == BP_NO. The way that I understand this is that need_static = 1 is the default and that returns the values from the info table in battery.c. If we can't talk to the battery, then it will use the table entry associated with DEFAULT_BATTERY_TYPE. It seems to me that when the battery is not present, need_static should be 1. That's why I want to understand better what the problem is in this situation.

Patch-set: 1
Reviewer: Gerrit User 1123814 <1123814@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>