Fuzz: Add support for fuzzing targets.
This adds the fuzz/ subfolder modeled after the test subfolder.
Fuzzing targets are added in similar fashion to unit tests. A
special build environment is needed to build the fuzz targets,
as they are intended for use with oss-fuzz.

The initial target included in this commit fuzzes part of the
PinWeaver vendor specific command for Cr50. More work is needed
to expand the coverage farther. Libprotobuf-mutator is used to
fuzz at a higher level than raw input, so valid operations can
be performed without the fuzzer needing to brute force hmacs,
hashes, and encryption keys.

BRANCH=none
BUG=None
TEST=make -j buildall

Change-Id: I877b7535374dd1a283f2eee1b076000d1c5fc4eb
Signed-off-by: Allen Webb <allenwebb@google.com>