waddledoo: don't enable pull-ups with low-voltage mode
It is not valid on the NPCX7 to simultaneously use low-voltage mode and
enable the pull-up, so pull-ups were being ignored.  This caused
spurious volume up and volume down button presses.

Prochot is similarly misconfigured and really should be an ADC input,
but for now just remove the pull-up.

These changes have been updated on the EC pinout table as well.

BUG=None
BRANCH=None
TEST=press buttons on waddledoo

Change-Id: I8fd6e13d0e0ca79dbb777563443da0858ac62187