Update patch set 2
Patch Set 2:

why is it necessary to move nvmem_init and run fips tests so early in the boot up process?

If this patch is accepted: who is going to re-enable console and take the AP out or reset?

My understanding was that it was important to suppress console output only to prevent showing FIPS initialization related output text, is this not the case?

Patch-set: 2