usb-pd: svdm_dp_attention used bit & instead of logical &&
The code works the same but it is not clear why the &&
was not chosen and no comments to give a reason. Since
both variables are always 0 or 1, the code will do
exactly the same thing but this is really what is
happening when you think about the logic behind this
comparison.

BUG=none
BRANCH=none
TEST=make buildall

Signed-off-by: Denis Brockus <dbrockus@google.com>
Change-Id: I4daf780e370978b468a5d1c22536b904a591e089