kracko: Move C1 SM5803 processing to the PD_INT task
Since the SM5803 shares an interrupt line with the TCPC, allow the
PD_INT task for C1 to process interrupts for this chip.  This will
ensure that any interrupts from the charger are handled at a high
priority and cannot leave the shared IRQ line low for extended periods.
(This CL aligns with CL:2803462)

BUG=none
BRANCH=dedede
TEST=On kracko, confirm charger attach to C1 works reliably.

Signed-off-by: Tommy Chung <tommy.chung@quanta.corp-partner.google.com>
Change-Id: I21622c1ca06db9b600079487053f6efdaa8cf88e