test: Remove TEST_CHECK macro
The TEST_CHECK macro returns a value (from the calling function) on
either success or failure. This differes from all other TEST_* macros in
test_util.h, which only return on failure. Returning on failure is
appropriate and allows short circuiting a test on the first failure but
returning on success results in short circuiting after the first
successful check and bypassing subsequent checks. This behavior
is somewhat confusing and easy to miss during review.

BRANCH=none
BUG=b:234781655
TEST=make runhosttests

Signed-off-by: Bobby Casey <bobbycasey@google.com>
Change-Id: I3777b427aa5e20a91689f86fc37daadffacf27f3