nct38xx-gpio-alert: Delete unused node
Due to upstream bug
    https://github.com/zephyrproject-rtos/zephyr/issues/61436
the nct38xx-gpio-alert driver is never enabled.

However, the downstream TCPC driver manually triggers the interrupt
handler of the nct38xx-gpio driver.  This manual trigger is not
compatible with the upstream driver, so delete the nct38xx-gpio-alert
node from all boards.

This change can be reverted once b:295587630 is fixed.

BUG=b:295587630
TEST=zmake compare-builds -a

Change-Id: I5541634ff0da2403369a02b1cd3b4a695a134d09