board/hyperdebug: Allow ST-link debugging by default
When declarations for pins of ports CN11/12 was added, the ST-link
debugging support was accidentally lost, as the two pins which are
connected to the onboard debugger at the top end of the HyperDebug board
were configured as GPIO inputs, rather than their default "alternate"
mode.

With this change, users of HyperDebug can still explicitly set the two
pins into GPIO mode if they want, but at power-up they will allow
debugging of the STM32L5 MCU on HyperDebug.

BUG=b:339034638
TEST=make BOARD=hyperdebug  (observing openocd making connection)

Change-Id: Icd136d18d9ca6a2f89da30e294ddabc2bc79f1db