[Do not merge] usb_pd: Reduce the frequenency of CC state polling
Currently the EC polls the CC state from TCPC eveytime when
tc_attached_src_run is called to detect SRC.Open state on the
monitored CC pin, this result in frequent I2C access to the TCPC.
This patch uses the timer to reduce the frequency of the CC state
polling. With this patch, the EC polls the CC state every
PD_T_SRC_DISCONNECT (15ms) when the Type-C state is in
Attached.SRC state.

Change-Id: I24e25b9d39cec86967ba6c5091ef25300254dd43